+++ packages/import/client/views/panels.js
@@ -7,3 +7,19 @@ Template.formImport.helpers({
 		}
 	}
 });
+
+Template.item_place_imported.onRendered(function() {
+	
+	var self = this,
+		tmpl = Template.instance();
+
+	self.$('.btn-editdel').btsConfirmButton(function(e) {
+		
+		Meteor.call('removePlace', self.data.id, function(err,res) {
+		console.log(self.data.id, err,res)
+			K.Map.removeItem(self.data);
+		
+			//Router.go('root');
+		});			
+	});
+});