+++ packages/import/collections/places.js
@@ -2,7 +2,7 @@
 K.extend({
 	findPlacesImportByUserId: function(userId) {
 		return Places.find({
-			'import': {$exists:true},
+			'import.name': {$exists:true},
 			'userId': userId
 		}, _.deepExtend({}, K.filters.placeItem, {
 				sort: { createdAt: -1 },