+++ packages/import/package.js
@@ -10,6 +10,7 @@ Package.describe({
 Package.onUse(function(api) {
   api.use([
     'keplerjs:core@'+version,
+    'keplerjs:edit@'+version,    
     'keplerjs:admin@'+version,
     'keplerjs:upload@'+version,
   ]);