+++ packages/photos/i18n/en.js
@@ -1,5 +1,6 @@
 
 i18n.map('en', {
+	label_addphotos: "Add Photos",
 	btn_avatar: "Photo profile",
 	upload_error_imageNotValid: "Only images JPG/PNG",
 });