+++ packages/photos/server/photos.js
@@ -1,11 +1,24 @@
 
 var fs = Npm.require('fs');
+var Future = Npm.require("fibers/future");
+var Imagemagick = Npm.require("imagemagick");
 
-var exifreader = Npm.require('exif-reader');
+if(!Imagemagick) {
+	console.warn('Photos: Imagemagick not installed');
+	throw new Meteor.Error(500);
+}
 
-//TODO require  npm Imagemagick
-// 
 var CHMOD = 0755;
+
+var ImagemagickSync = {};
+["identify", "readMetadata", "convert", "resize", "crop"].forEach(function(fn) {
+    ImagemagickSync[fn] = function(args) {
+        var future = new Future();
+        var cb = future.resolver();
+        Imagemagick[fn](args, cb);
+        return future.wait();
+    };
+});    
 
 Meteor.methods({
 	
@@ -18,13 +31,6 @@ Meteor.methods({
 			username = Meteor.user().username,
 			fileName = K.Util.sanitize.filename( username +'_'+ K.Util.time() ),
 			fileOut = fileName + '.ori.jpg';
-
-		if(!Imagemagick)
-		{
-			console.log('Photos: error Imagemagick not installed');
-			throw new Meteor.Error(500);
-			return null;
-		}
 
 		fs.writeFileSync(sets.path + fileOut, fileObj.blob, 'binary');
 		fs.chmodSync(sets.path + fileOut, CHMOD);
@@ -45,8 +51,8 @@ Meteor.methods({
 		console.log('Photos: resizing...');
 
 		try {
-			Imagemagick.crop(imgOpts);
-			fs.chmodSync(basePath + fileMin, CHMOD);
+			ImagemagickSync.crop(imgOpts);
+			fs.chmodSync(imgOpts.dstPath, CHMOD);
 		}
 		catch(e) {
 			console.log('Photos: error ', e);
@@ -67,10 +73,55 @@ Meteor.methods({
 
 	exifPhoto: function(fileObj, sets) {
 		
-		var meta = exifreader(fileObj.blob);
+		if(!this.userId) return null;
 
-		console.log('Photos: exifPhoto meta', meta)
+		var basePath = sets.path,
+			baseUrl = sets.url,
+			username = Meteor.user().username,
+			fileName = K.Util.sanitize.filename( username +'_'+ K.Util.time() ),
+			fileOut = fileName + '.ori.jpg';
+
+		fs.writeFileSync(sets.path + fileOut, fileObj.blob, 'binary');
+		fs.chmodSync(sets.path + fileOut, CHMOD);
+
+		try {
+			var meta = ImagemagickSync.readMetadata(basePath + fileOut);
+		}
+		catch(e) {
+			console.log('Photos: error ', e);
+			throw new Meteor.Error(500, i18n('upload_error_imageNotValid') );
+			return i18n('upload_error_imageNotValid');
+		}
+
+		var exif = _.omit(meta.exif,'makerNote');
+
+		console.log('Photos: exifPhoto ', exif)
+
+		/*gpsAltitude: "447000/1000"
+		gpsAltitudeRef: "0"
+		gpsDateStamp: "2018:11:10"
+		gpsInfo: "20990"
+		gpsLatitude: "46/1, 3/1, 51821/1000"
+		gpsLatitudeRef: "N"
+		gpsLongitude: "11/1, 9/1, 29574/1000"
+		gpsLongitudeRef: "E"
+		gpsMapDatum: "WGS-84"
+		gpsStatus: "A"
+		gpsTimeStamp: "10/1, 42/1, 25000/1000"
+		gpsVersionID: "2, 2, 0, 0"*/
+
+		var lats = exif.gpsLatitude.split(',').map(function(v) {
+				return parseInt(v.split('/')[0]);
+			}),
+			lngs = exif.gpsLongitude.split(',').map(function(v) {
+				return parseInt(v.split('/')[0]);
+			}),
+			lat = lats[0]+'°'+lats[1]+"'"+(lats[2]/1000)+'"'+exif.gpsLatitudeRef,
+			lng = lngs[0]+'°'+lngs[1]+"'"+(lngs[2]/1000)+'"'+exif.gpsLongitudeRef;			
+//example: parseDMS('59°12\'7.7"N 02°15\'39.6"W')
+
+		exif.loc = K.Util.geo.parseDMS(lat+' '+lng);
 		
-		return meta;
+		return exif;
 	}
 });