+++ packages/photos/server/places.js
@@ -3,6 +3,14 @@
 Meteor.methods({
 	
 	insertPlacePhotos: function(fileObj, sets) {
-		return Meteor.call('exifPhoto', fileObj, sets);
+
+		var url = Meteor.call('resizePhoto', fileObj, sets);
+
+		var exif = Meteor.call('exifPhoto', fileObj, sets);
+
+		return {
+			url: url,
+			loc: exif.loc
+		};
 	}
 });