+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
-...
+- plugin import remove button in places list
+
 
 ## v1.5.4
 - new plugin photos